requirements.txt
apiclient==1.0.2
dj-config-url==0.1.1
Django==1.9.5
django-filter==0.11.0
djangorestframework==3.3.3
google-api-python-client==1.5.0
gunicorn==19.4.5
httplib2==0.9.2
Markdown==2.6.6
oauth2client==2.0.2
psycopg2==2.6.1
pyasn1==0.1.9
pyasn1-modules==0.0.8
rsa==3.4.2
simplejson==3.8.2
six==1.10.0
uritemplate==0.6
urllib3==1.15.1
whitenoise==3.0
